// apollo-shit/generated/server-graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type AdditionalEntityFields = {
  path?: Maybe<Scalars['String']>;
  type?: Maybe<Scalars['String']>;
};

export type BaseOrder = {
  __typename?: 'BaseOrder';
  _id: Scalars['String'];
  active: Scalars['Boolean'];
  items?: Maybe<Array<Maybe<Item>>>;
};

export type ErrorResponse = {
  __typename?: 'ErrorResponse';
  code?: Maybe<Scalars['String']>;
  httpStatusCode?: Maybe<Scalars['Int']>;
  message: Scalars['String'];
};

export type Item = {
  __typename?: 'Item';
  _id: Scalars['String'];
  isFavorite?: Maybe<Scalars['Boolean']>;
  name: Scalars['String'];
  productImageUrl?: Maybe<Scalars['String']>;
  productUrl: Scalars['String'];
};